# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "后台管理系统 API 文档",
        "title": "Kingfisher Admin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/docs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "目录内文档分页列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目录 ID",
                        "name": "dir_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文档列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/kingfisher_core_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/kingfisher_core_response.PageData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "创建文档（初始 draft，版本 1）",
                "responses": {}
            }
        },
        "/docs/:id": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "文档详情",
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "更新文档（追加新版本）",
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "删除文档（级联删版本）",
                "responses": {}
            }
        },
        "/docs/:id/publish": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "发布文档 draft → published",
                "responses": {}
            }
        },
        "/docs/:id/restore": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "还原到指定版本（追加新版本）",
                "responses": {}
            }
        },
        "/docs/:id/unpublish": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "撤稿 published → draft",
                "responses": {}
            }
        },
        "/docs/:id/versions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "文档版本历史列表",
                "responses": {}
            }
        },
        "/docs/:id/versions/:no": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "查看指定版本内容",
                "responses": {}
            }
        },
        "/docs/dirs": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "创建目录",
                "responses": {}
            }
        },
        "/docs/dirs/:id": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "更新目录（改名/移动/排序）",
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "删除目录（有子目录/文档则拒绝）",
                "responses": {}
            }
        },
        "/docs/tree": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "当前用户可见目录树",
                "responses": {
                    "200": {
                        "description": "目录树",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/kingfisher_core_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/kingfisher_extends_doc_domain.DocDirectory"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/docs/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Doc"
                ],
                "summary": "上传文档图片",
                "parameters": [
                    {
                        "type": "file",
                        "description": "图片文件（png/jpg/jpeg/gif/webp，≤2MB）",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "url",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/kingfisher_core_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "kingfisher_core_response.PageData": {
            "type": "object",
            "properties": {
                "items": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "kingfisher_core_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "kingfisher_extends_doc_domain.DocDirectory": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/kingfisher_extends_doc_domain.DocDirectory"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "docs": {
                    "description": "该目录下当前用户可见的文档（叶子节点）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/kingfisher_extends_doc_domain.DocTreeItem"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                },
                "visibility": {
                    "description": "shared | private",
                    "type": "string"
                }
            }
        },
        "kingfisher_extends_doc_domain.DocTreeItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "visibility": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}